requirements.txt
spidev
py-opc

